Add another config test dir and file (dir5 and file5)
Test that a complex pattern of excludes and alwaysincludes allows us
always include a file or dir buried inside an excluded dir.

Test that the correct sequence is added when we try to do this using the 
GUI.

Test that deselecting the AlwaysIncluded dir only removes its own node, 
not the parents.


git-svn-id: https://svn.code.sf.net/p/boxi/code/trunk/boxi@576 bde4d9af-c013-0410-a270-fa18cafdf79f